refactor: Separate TableViewHeader  from TableView and put on top of AttendanceViewController
